retry crates.io requests when requesting crate-details
